[REFACTOR] Turn createAPIMonitor APP into component
