+++ src/components/APIMonitorEditor/optionHelper.js
@@ -1,0 +1,59 @@
+const methodOption = [
+    {
+        key: "GET",
+        value: "GET",
+    },
+    {
+        key: "POST",
+        value: "POST",
+    },
+    {
+        key: "PUT",
+        value: "PUT",
+    },
+    {
+        key: "PATCH",
+        value: "PATCH",
+    },
+    {
+        key: "DELETE",
+        value: "DELETE",
+    },
+  ]
+
+  const intervalOption = [
+    {
+        key: "1MIN",
+        value: "1 Minute",
+    },
+    {
+        key: "2MIN",
+        value: "2 Minutes",
+    },
+    {
+        key: "3MIN",
+        value: "3 Minutes",
+    },
+    {
+        key: "5MIN",
+        value: "5 Minutes",
+    },
+    {
+        key: "10MIN",
+        value: "10 Minutes",
+    },
+    {
+        key: "15MIN",
+        value: "15 Minutes",
+    },
+    {
+        key: "30MIN",
+        value: "30 Minutes",
+    },
+    {
+        key: "60MIN",
+        value: "60 Minutes",
+    },
+]
+
+export {methodOption, intervalOption};
